Modify: 중복태그제거, 필요없는 스타일 제거, 헤더높이 지정 태그 변경
