+++ src/components/Header.scss
@@ -1,0 +1,160 @@
+.header {
+  .header-wrap {
+    box-sizing: border-box;
+    max-width: 1720px;
+    height: 100%;
+    margin: 0 48px;
+    background: #fff;
+
+    .inner {
+      display: grid;
+      grid-column-gap: 20px;
+      grid-template-columns: repeat(13, minmax(0, 1fr));
+      height: 90px;
+
+      .inner-menu-container {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        pointer-events: all;
+
+        .menu-icon {
+          width: 24px;
+          height: 24px;
+          margin-bottom: 10px;
+          cursor: pointer;
+          &:hover {
+            background-color: yellow;
+          }
+        }
+      }
+
+      .inner-main {
+        grid-column: 2/14;
+        grid-row: 1;
+        border-bottom: 1px solid #dbdbdb;
+
+        .logo {
+          display: flex;
+          align-items: center;
+          justify-content: space-between;
+          height: 100%;
+
+          .logo-img {
+            width: 90px;
+            height: 36px;
+          }
+          .search {
+            display: block;
+            position: relative;
+            width: 85%;
+
+            span {
+              position: absolute;
+              top: 17px;
+              left: 15px;
+              color: grey;
+            }
+
+            .search-input {
+              width: 100%;
+              margin: 0 auto;
+              padding: 0.8125rem;
+              padding-inline-start: 3rem;
+              padding-inline-end: 3.6rem;
+              border: 2px solid transparent;
+              border-radius: 50em;
+              outline: 0;
+              background-color: lightyellow;
+              &:hover {
+                background-color: yellow;
+                color: black;
+              }
+            }
+          }
+
+          .user-img,
+          .login-img,
+          .cart-img {
+            width: 24px;
+            height: 24px;
+            cursor: pointer;
+            &:hover {
+              background-color: yellow;
+            }
+          }
+        }
+      }
+    }
+  }
+
+  .nav-wrap {
+    box-sizing: border-box;
+    max-width: 1720px;
+    height: 100%;
+    margin: 0 48px;
+    border-bottom: 1px solid black;
+    border-bottom: none;
+    background: #fff;
+    .nav-inner {
+      display: grid;
+      grid-column-gap: 20px;
+      grid-template-columns: repeat(13, minmax(0, 1fr));
+
+      .nav-main {
+        display: flex;
+        grid-column: 2/14;
+        grid-row: 1;
+        flex: 0 0 100%;
+        flex-direction: row;
+        margin-left: 0;
+        padding-left: 0;
+        .nav-left-item {
+          display: flex;
+          width: 100%;
+          li {
+            a {
+              display: block;
+              padding: 0.625rem 0.9375rem;
+              color: black;
+              font-size: 0.875rem;
+              font-weight: 700;
+              text-decoration: none;
+            }
+
+            .nav-left-list-first {
+              padding-left: 0;
+            }
+          }
+        }
+
+        .nav-right-item {
+          display: flex;
+          flex: 0 1 100%;
+          align-items: center;
+          justify-content: flex-end;
+          width: auto;
+
+          .nav-right-list {
+            display: flex;
+            align-items: center;
+            .right-icon {
+              display: flex;
+              width: 16px;
+              height: 16px;
+            }
+            .nav-right-items {
+              display: block;
+              padding: 10px 0px 10px 30px;
+              color: black;
+              font-size: 0.875rem;
+              font-weight: 700;
+              text-decoration: none;
+            }
+          }
+        }
+      }
+    }
+  }
+}
